lat and lon are float

+++ src/trace.py
@@ -19,7 +19,7 @@ cache_name = "ip_loc_cache.txt"
 
 class Trace:
     def __init__(self):
-        self.ip_locations: Dict[str, Tuple[int, int]] = {}
+        self.ip_locations: Dict[str, Tuple[float, float]] = {}
         self.blacklisted_ips: Set[str] = set()
 
     def read_from_file(self) -> None:
@@ -28,7 +28,7 @@ class Trace:
                 with open(cache_name, 'r') as cache:
                     for line in cache.readlines():
                         ip, lat, lon = line.split(',')
-                        self.ip_locations[ip] = lat, lon
+                        self.ip_locations[ip] = float(lat), float(lon)
             except Exception as e:
                 logging.error(f'Unable to load cache: {e}')
